+++ Examples/+BrDemo/AFC_5_1_Falsification.m
@@ -13,7 +13,6 @@ BrAFC
 %% 
 % First we create the parameter search domain and load specifications. 
 AFC_Falsify = BrAFC.copy();
-
 AFC_Falsify.SetParamRanges({'Pedal_Angle_pulse_period', 'Pedal_Angle_pulse_amp'}, [10 20; 10 60]);
 STL_ReadFile('AFC_simple_spec.stl');