+++ InstallBreach.m
@@ -2,6 +2,8 @@ function InstallBreach(varargin)
 %INSTALLBREACH compiles C/C++ mex functions used by Breach. Needs only to
 % be called once after an update of Breach
 
+InitBreach
+
 global BreachGlobOpt;
 options.verbose = 0;
 options.linear_interp = 0; 
@@ -23,12 +25,10 @@ if silent
    FLAGS = [FLAGS ' -silent '];
 end
 
-InitBreach
 cdr = pwd;
 dr = which('InstallBreach');
 
 % setup directories of interest
-
 breach_dir = dr(1:end-16);
 breach_src_dir = [breach_dir filesep 'Core' filesep 'src']; 
 stl_src_dir  = [breach_dir filesep '@STL_Formula' filesep 'private' filesep 'src'];